Refactor: 로딩 lib로 이동
